Update `Migrations.sol` to use `uint` instead of `uint256` for `last_completed_migration`
Update `truffle-config.js` to remove compiler version configuration

Update `nohup.out` with npm warnings and ganache-cli installation details

+++ contracts/Migrations.sol
@@ -2,7 +2,7 @@ pragma solidity ^0.8.0;
 
 contract Migrations {
     address public owner;
-    uint256 public last_completed_migration;
+    uint public last_completed_migration;
 
     constructor() {
         owner = msg.sender;
@@ -13,7 +13,7 @@ contract Migrations {
         _;
     }
 
-    function setCompleted(uint256 completed) public restricted {
+    function setCompleted(uint completed) public restricted {
         last_completed_migration = completed;
     }
 }
